Start on a YAML indentexpr plugin
To write a Lua indent function, one must set indentexpr to a Vimscript
function that calls the Lua function. I don't know if there's any real
advantage to writing the function in Lua, apart from reducing how much
Vimscript code I have to write, but it works fine.

It's possible that the calls back-and-forth from Vimscript to Lua (and
back, because the Lua code calls vim.fn.xxx) will make it slower, but
who cares?

Right now, the plugin just faithfully reproduces the indentation of the
previous line, so the effect is like autoindent, with no YAML-specific
smarts at all. That's still better than the built-in indent plugin,
though.